Fiddle with zsh prompt some more
Based on the customized agnoster.zsh-theme at
https://gist.github.com/smileart/3750104.

There are lots of annoying Unicode issues. Just to get the file to
save, I had to switch to raw UTF-8. And then the extra characters I'm
trying to use wouldn't display in my terminal.

And the basic way it handles the git rprompt isn't meshing
correctly with what I already have.

It would be nice to add more color changes to make it more dynamic and
flashier (well, it seems that way before I've had a chance to try it),
I don't know if it's worth the time/effort I've already sunk.

Still, keeping this around as an experiment, in case I decide it's
worth playing with more later.